[ADD] Lights can now lighten OR darken sectors.

+++ includes/render.h
@@ -6,7 +6,7 @@
 /*   By: gaerhard <gaerhard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/09 13:20:37 by lnicosia          #+#    #+#             */
-/*   Updated: 2019/09/05 11:35:04 by lnicosia         ###   ########.fr       */
+/*   Updated: 2019/09/05 13:21:48 by lnicosia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -113,7 +113,7 @@ typedef struct		s_render
 	int				texture;
 	int				floor_texture;
 	int				ceiling_texture;
-	uint8_t			brightness;
+	short			brightness;
 	Uint32			light_color;
 	double			x1z1;
 	double			x2z2;
@@ -155,7 +155,7 @@ typedef struct		s_render_object
 	int				xend;
 	int				yend;
 	int				index;
-	uint8_t			brightness;
+	short			brightness;
 	Uint32			light_color;
 	double			xrange;
 	double			yrange;
